\name{internal functions not to be used by users directly}
\alias{bsplinepen.fda}
\title{internal functions not to be used by users directly}
\description{
  Internal Functions Not to Be Used by Users Directly.
}
\details{
  These are not to be called by the user, and are likely to be defunct in future versions.
}
\keyword{ internal }
