\name{hisemi-package}
\alias{hisemi-package}
\alias{hisemi}
\docType{package}
\title{
The package includes functions for fitting hierarchical semiparametric regression model to a large number of parametric test statistics. 
}
\description{
The package includes functions for fitting hierarchical semiparametric regression model to a large number of parametric test statistics
}
\details{
\tabular{ll}{
Package: \tab hisemi\cr
Type: \tab Package\cr
Version: \tab 1.0-123\cr
Date: \tab 2010-09-17\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
\code{\link{penLik.EMNewton}} is the major interface function to be called. 
}
\author{
Long Qu \email{longor@iastate.edu}, Dan Nettleton \email{dnett@iastate.edu}, Jack C.M. Dekkers \email{jdekkers@iastate.edu}
Maintainer: Long Qu \email{longor@iastate.edu}
}
\references{ Long Qu, Dan Nettleton, Jack Dekkers. A hierarchical semiparametric model for incorporating inter-gene relationship information for analysis of genomic data. manuscript.  }
\keyword{ package }
\keyword{ models }
\keyword{ regression }
\keyword{ optimize }
\seealso{
\code{\link[pi0]{pi0}}
}
\examples{
}
